testing/web-platform/tests/encoding-detection/fa-windows-1256.tentative.html

<!doctype html>
<title>fa windows-1256</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>�&#1740;� &#1740;� ��� ��Ґ���&#1740; ���ǘ�� ���.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1256", 'Expected windows-1256');
    }, "Check detection result");
      done();
};
</script>
